changed parsing lat & lng from int to float

+++ src/features/map/mapSlice.ts
@@ -38,8 +38,8 @@ export const mapSlice = createSlice({
         mapCenterChanged: (state, {payload: {coordinates}}: PayloadAction<{coordinates: string}>) => {            
             try {
                 const [latString, lngString] = coordinates.split(',');
-                const lat = parseInt(latString);                
-                const lng = parseInt(lngString);
+                const lat = parseFloat(latString);                
+                const lng = parseFloat(lngString);
                 state.mapCenter = {lat,lng}
             } catch(e) {
                 console.log(e);
